move feed settings from database to instance settings, other small cleanup and fixes
